FAC-834 Move attribute insight related class to 'attributeinsight' pkg
