+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/attributeinsight/UpdateAttributeInsightRequestDto.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.adapter.in.rest.attribute;
+package org.flickit.assessment.core.adapter.in.rest.attributeinsight;
 
 public record UpdateAttributeInsightRequestDto(String assessorInsight) {
 }
