+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/port/in/attributeinsight/GetAttributeInsightUseCase.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.port.in.attribute;
+package org.flickit.assessment.core.application.port.in.attributeinsight;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.EqualsAndHashCode;
