+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/attributeinsight/ApproveAttributeInsightService.java
@@ -1,10 +1,10 @@
-package org.flickit.assessment.core.application.service.attribute;
+package org.flickit.assessment.core.application.service.attributeinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentPermission;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.core.application.port.in.attribute.ApproveAttributeInsightUseCase;
+import org.flickit.assessment.core.application.port.in.attributeinsight.ApproveAttributeInsightUseCase;
 import org.flickit.assessment.core.application.port.out.attributeinsight.ApproveAttributeInsightPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
