+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/attributeinsight/ApproveAttributeInsightServiceTest.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.application.service.attribute;
+package org.flickit.assessment.core.application.service.attributeinsight;
 
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentPermission;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.core.application.port.in.attribute.ApproveAttributeInsightUseCase;
+import org.flickit.assessment.core.application.port.in.attributeinsight.ApproveAttributeInsightUseCase;
 import org.flickit.assessment.core.application.port.out.attributeinsight.ApproveAttributeInsightPort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
